requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
boto3==1.34.138 
python-dotenv==1.0.0
sqlalchemy==2.0.31 
alembic==1.13.2
databases==0.9.0
asyncpg
psycopg2-binary
pytest==8.2.2 